+++ npm/vite-dev-server/cypress/components/large-third-party-dep.spec.ts
@@ -1,0 +1,7 @@
+import _ from 'lodash'
+
+describe('Large 3rd party library without tree-shaking', () => {
+  it('successfully imports lodash', () => {
+    expect(_.isBoolean(true)).to.be.true
+  })
+})