+++ packages/runner/src/lib/event-manager.js
@@ -3,18 +3,18 @@ import { EventEmitter } from 'events'
 import Promise from 'bluebird'
 import { action } from 'mobx'
 
-import { client, circularParser } from '@packages/socket/lib/browser'
+import { client } from '@packages/socket'
 
 import automation from './automation'
 import logger from './logger'
 import studioRecorder from '../studio/studio-recorder'
+import selectorPlaygroundModel from '../selector-playground/selector-playground-model'
 
 import $Cypress, { $ } from '@packages/driver'
 
 const ws = client.connect({
   path: '/__socket.io',
   transports: ['websocket'],
-  parser: circularParser,
 })
 
 ws.on('connect', () => {
@@ -26,8 +26,8 @@ const driverToLocalAndReporterEvents = 'run:start run:end'.split(' ')
 const driverToSocketEvents = 'backend:request automation:request mocha recorder:frame'.split(' ')
 const driverTestEvents = 'test:before:run:async test:after:run'.split(' ')
 const driverToLocalEvents = 'viewport:changed config stop url:changed page:loading visit:failed'.split(' ')
-const socketRerunEvents = 'runner:restart'.split(' ')
-const socketToDriverEvents = 'net:event'.split(' ')
+const socketRerunEvents = 'runner:restart watched:file:changed'.split(' ')
+const socketToDriverEvents = 'net:event script:error'.split(' ')
 const localToReporterEvents = 'reporter:log:add reporter:log:state:changed reporter:log:remove'.split(' ')
 
 const localBus = new EventEmitter()
@@ -82,6 +82,12 @@ const eventManager = {
       switch (msg) {
         case 'change:cookie':
           Cypress.Cookies.log(data.message, data.cookie, data.removed)
+          break
+        case 'create:download':
+          Cypress.downloads.start(data)
+          break
+        case 'complete:download':
+          Cypress.downloads.end(data)
           break
         default:
           break
@@ -382,9 +388,7 @@ const eventManager = {
     Cypress.on('log:added', (log) => {
       const displayProps = Cypress.runner.getDisplayPropsForLog(log)
 
-      if (studioRecorder.isActive) {
-        displayProps.hookId = studioRecorder.hookId
-      }
+      this._interceptStudio(displayProps)
 
       reporterBus.emit('reporter:log:add', displayProps)
     })
@@ -392,9 +396,7 @@ const eventManager = {
     Cypress.on('log:changed', (log) => {
       const displayProps = Cypress.runner.getDisplayPropsForLog(log)
 
-      if (studioRecorder.isActive) {
-        displayProps.hookId = studioRecorder.hookId
-      }
+      this._interceptStudio(displayProps)
 
       reporterBus.emit('reporter:log:state:changed', displayProps)
     })
@@ -461,6 +463,12 @@ const eventManager = {
         studioRecorder.setFileDetails(test.invocationDetails)
       }
     })
+
+    Cypress.on('test:after:run', (test) => {
+      if (studioRecorder.isOpen && test.state !== 'passed') {
+        studioRecorder.testFailed()
+      }
+    })
   },
 
   _runDriver (state) {
@@ -497,6 +505,7 @@ const eventManager = {
     Cypress.stop()
 
     studioRecorder.setInactive()
+    selectorPlaygroundModel.setOpen(false)
 
     return this._restart()
     .then(() => {
@@ -541,6 +550,19 @@ const eventManager = {
         Cypress.runner.setOnlyTestId(studioRecorder.testId)
       }
     }
+  },
+
+  _interceptStudio (displayProps) {
+    if (studioRecorder.isActive) {
+      displayProps.hookId = studioRecorder.hookId
+
+      if (displayProps.name === 'visit' && displayProps.state === 'failed') {
+        studioRecorder.testFailed()
+        reporterBus.emit('test:set:state', studioRecorder.testError, _.noop)
+      }
+    }
+
+    return displayProps
   },
 
   emit (event, ...args) {