+++ packages/runner/src/studio/studio-recorder.js
@@ -1,5 +1,6 @@
 import { action, computed, observable } from 'mobx'
 import { $ } from '@packages/driver'
+import $driverUtils from '@packages/driver/src/cypress/utils'
 
 import eventManager from '../lib/event-manager'
 
@@ -19,7 +20,12 @@ const eventTypes = [
 const eventsWithValue = [
   'change',
   'keydown',
-  'select',
+]
+
+const internalMouseEvents = [
+  'mousedown',
+  'mouseover',
+  'mouseout',
 ]
 
 export class StudioRecorder {
@@ -31,11 +37,12 @@ export class StudioRecorder {
   @observable isLoading = false
   @observable isActive = false
   @observable url = null
+  @observable isFailed = false
   @observable _hasStarted = false
-  @observable _preventUrlInput = false
 
   fileDetails = null
   _currentId = 1
+  _previousMouseEvent = null
 
   @computed get hasRunnableId () {
     return !!this.testId || !!this.suiteId
@@ -50,7 +57,7 @@ export class StudioRecorder {
   }
 
   @computed get isReady () {
-    return this.isOpen && this.isEmpty && !this.isLoading
+    return this.isOpen && this.isEmpty && !this.isLoading && !this.isFailed
   }
 
   @computed get hookId () {
@@ -58,7 +65,14 @@ export class StudioRecorder {
   }
 
   @computed get needsUrl () {
-    return this.isActive && !this.url && !this._preventUrlInput
+    return this.isActive && !this.url && !this.isFailed
+  }
+
+  @computed get testError () {
+    return {
+      id: this.testId,
+      state: 'failed',
+    }
   }
 
   @computed get saveError () {
@@ -70,6 +84,10 @@ export class StudioRecorder {
         docsUrl: 'https://on.cypress.io/studio-beta',
       },
     }
+  }
+
+  get Cypress () {
+    return eventManager.getCypress()
   }
 
   @action setTestId = (testId) => {
@@ -114,12 +132,20 @@ export class StudioRecorder {
     this.url = url
   }
 
-  @action preventUrlInput = () => {
-    this._preventUrlInput = true
+  @action testFailed = () => {
+    this.isFailed = true
   }
 
   setFileDetails = (fileDetails) => {
     this.fileDetails = fileDetails
+  }
+
+  _clearPreviousMouseEvent = () => {
+    this._previousMouseEvent = null
+  }
+
+  _matchPreviousMouseEvent = (el) => {
+    return this._previousMouseEvent && $(el).is(this._previousMouseEvent.element)
   }
 
   @action start = (body) => {
@@ -150,7 +176,7 @@ export class StudioRecorder {
     this.url = null
     this._hasStarted = false
     this._currentId = 1
-    this._preventUrlInput = false
+    this.isFailed = false
   }
 
   @action cancel = () => {
@@ -181,9 +207,7 @@ export class StudioRecorder {
   @action visitUrl = (url = this.url) => {
     this.setUrl(url)
 
-    const Cypress = eventManager.getCypress()
-
-    Cypress.cy.visit(this.url)
+    this.Cypress.cy.visit(this.url)
 
     this.logs.push({
       id: this._getId(),
@@ -194,6 +218,8 @@ export class StudioRecorder {
   }
 
   attachListeners = (body) => {
+    if (this.isFailed) return
+
     this._body = body
 
     eventTypes.forEach((event) => {
@@ -202,6 +228,15 @@ export class StudioRecorder {
         passive: true,
       })
     })
+
+    internalMouseEvents.forEach((event) => {
+      this._body.addEventListener(event, this._recordMouseEvent, {
+        capture: true,
+        passive: true,
+      })
+    })
+
+    this._clearPreviousMouseEvent()
   }
 
   removeListeners = () => {
@@ -212,6 +247,39 @@ export class StudioRecorder {
         capture: true,
       })
     })
+
+    internalMouseEvents.forEach((event) => {
+      this._body.removeEventListener(event, this._recordMouseEvent, {
+        capture: true,
+      })
+    })
+
+    this._clearPreviousMouseEvent()
+  }
+
+  _trustEvent = (event) => {
+    // only capture events sent by the actual user
+    // but disable the check if we're in a test
+    return event.isTrusted || this.Cypress.env('INTERNAL_E2E_TESTS') === 1
+  }
+
+  _recordMouseEvent = (event) => {
+    if (!this._trustEvent(event)) return
+
+    const { type, target } = event
+
+    if (type === 'mouseout') {
+      return this._clearPreviousMouseEvent()
+    }
+
+    // we only replace the previous mouse event if the element is different
+    // since we want to use the oldest possible selector
+    if (!this._matchPreviousMouseEvent(target)) {
+      this._previousMouseEvent = {
+        element: target,
+        selector: this.Cypress.SelectorPlayground.getSelector($(target)),
+      }
+    }
   }
 
   _getId = () => {
@@ -320,17 +388,13 @@ export class StudioRecorder {
   _shouldRecordEvent = (event, $el) => {
     const tagName = $el.prop('tagName')
 
-    return !(tagName !== 'INPUT' && event.type === 'keydown')
+    return !((tagName !== 'INPUT' && event.type === 'keydown') ||
+      (tagName === 'SELECT' && event.type === 'click') ||
+      tagName === 'OPTION')
   }
 
   @action _recordEvent = (event) => {
-    const Cypress = eventManager.getCypress()
-
-    // only capture events sent by the actual user
-    // but disable the check if we're in an e2e test
-    if (!event.isTrusted && Cypress.env('INTERNAL_E2E_TESTS') !== 1) {
-      return
-    }
+    if (this.isFailed || !this._trustEvent(event)) return
 
     const $el = $(event.target)
 
@@ -338,21 +402,27 @@ export class StudioRecorder {
       return
     }
 
-    const selector = Cypress.SelectorPlayground.getSelector($el)
-
     const name = this._getName(event, $el)
     const message = this._getMessage(event, $el)
 
-    if (name === 'change') {
+    if (name === 'change' || (name === 'type' && !message)) {
       return
     }
+
+    let selector = ''
+
+    if (name === 'click' && this._matchPreviousMouseEvent($el)) {
+      selector = this._previousMouseEvent.selector
+    } else {
+      selector = this.Cypress.SelectorPlayground.getSelector($el)
+    }
+
+    this._clearPreviousMouseEvent()
 
     const filteredLog = this._filterLastLog(selector, name, message)
 
     if (filteredLog) {
-      this._updateLog(filteredLog)
-
-      return
+      return this._updateLog(filteredLog)
     }
 
     this._addLog({
@@ -380,7 +450,7 @@ export class StudioRecorder {
       testId: this.testId,
       hookId: this.hookId,
       name,
-      message,
+      message: message ? $driverUtils.stringifyActual(message) : null,
       type,
       state: 'passed',
       instrument: 'command',
@@ -447,6 +517,15 @@ export class StudioRecorder {
       // and the click event always comes first
       if (name === 'select' && lastLog.name === 'click') {
         lastLog.name = 'select'
+        lastLog.message = message
+
+        return lastLog
+      }
+
+      // Cypress automatically issues a .click before every type
+      // so we can filter out the extra click event
+      if (name === 'type' && lastLog.name === 'click') {
+        lastLog.name = 'type'
         lastLog.message = message
 
         return lastLog