+++ packages/server/lib/modes/run.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console, @cypress/dev/arrow-body-multiline-braces  */
 const _ = require('lodash')
+const { app } = require('electron')
 const la = require('lazy-ass')
 const pkg = require('@packages/root')
 const path = require('path')
@@ -11,13 +12,13 @@ const logSymbols = require('log-symbols')
 
 const recordMode = require('./record')
 const errors = require('../errors')
-const Project = require('../project')
+const { ProjectBase } = require('../project-base')
 const Reporter = require('../reporter')
 const browserUtils = require('../browsers')
 const openProject = require('../open_project')
 const videoCapture = require('../video_capture')
+const { fs } = require('../util/fs')
 const runEvents = require('../plugins/run_events')
-const fs = require('../util/fs')
 const env = require('../util/env')
 const trash = require('../util/trash')
 const random = require('../util/random')
@@ -27,7 +28,6 @@ const newlines = require('../util/newlines')
 const terminal = require('../util/terminal')
 const specsUtil = require('../util/specs')
 const humanTime = require('../util/human_time')
-const electronApp = require('../util/electron-app')
 const settings = require('../util/settings')
 const chromePolicyCheck = require('../util/chrome_policy_check')
 const experiments = require('../experiments')
@@ -604,7 +604,7 @@ const openProjectCreate = (projectRoot, socketId, args) => {
 const createAndOpenProject = function (socketId, options) {
   const { projectRoot, projectId } = options
 
-  return Project
+  return ProjectBase
   .ensureExists(projectRoot, options)
   .then(() => {
     // open this project without
@@ -1476,11 +1476,23 @@ module.exports = {
           }
 
           if (!specs.length) {
-            errors.throw('NO_SPECS_FOUND', config.integrationFolder, specPattern)
+            // did we use the spec pattern?
+            if (specPattern) {
+              errors.throw('NO_SPECS_FOUND', projectRoot, specPattern)
+            } else {
+              // else we looked in the integration folder
+              errors.throw('NO_SPECS_FOUND', config.integrationFolder, specPattern)
+            }
           }
 
           if (browser.family === 'chromium') {
             chromePolicyCheck.run(onWarning)
+          }
+
+          if (options.componentTesting) {
+            specs = specs.filter((spec) => {
+              return spec.specType === 'component'
+            })
           }
 
           const runAllSpecs = ({ beforeSpecRun, afterSpecRun, runUrl, parallel }) => {
@@ -1545,11 +1557,16 @@ module.exports = {
     })
   },
 
-  run (options) {
-    return electronApp
-    .waitForReady()
-    .then(() => {
-      return this.ready(options)
+  async run (options) {
+    // electron >= 5.0.0 will exit the app if all browserwindows are closed,
+    // this is obviously undesirable in run mode
+    // https://github.com/cypress-io/cypress/pull/4720#issuecomment-514316695
+    app.on('window-all-closed', () => {
+      debug('all BrowserWindows closed, not exiting')
     })
+
+    await app.whenReady()
+
+    return this.ready(options)
   },
 }