+++ packages/server/test/e2e/0_reporters_spec.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const fs = require('../../lib/util/fs')
+const { fs } = require('../../lib/util/fs')
 const glob = require('../../lib/util/glob')
 const e2e = require('../support/helpers/e2e').default
 const Fixtures = require('../support/helpers/fixtures')